Simplify and unify query language
There are two important changes here. The first, and most important, is that a single query type is defined for any operations that act on datasets or the records that they contain. This unifies the PifQuery and DatasetQuery operations. The new class is io.citrine.jcc.search.core.query.BaseDatasetQuery. The derived classes, which replace PifQuery and DatasetQuery, are PifReturningQuery and DatasetReturningQuery. These new classes do not define any new filtering capabilities, but just contain additional formatting or metadata options specific to the types of records that they return.

The second change is that the method names and types are unified throughout the package. There were some differencens in the way that getters, setters, and other similar methods were named between the query language and results objects.

All of the changes here are meant to be backwards compatible _through the serialization/deserialization process_. This means that requests made to the API from old versions of this library will still execute correctly. However, users of the java code will need to update their own codes if they upgrade JCC versions to the 2.x series.