Add path to extracted values to PIF search hits
The extracted map in a PIF search hit stores values that are pulled from any location within a PIF during a query. This adds a peer to that field that stores the paths to the values that were extracted. That path is stored as a string in JSON pointer notation. There is a new parameter in PifSystemReturningQuery that toggles on the return of the paths of extractions.