Add simple search fields to all query objects
The simple search field is meant to store a string that will be tokenized and filtered over all fields in the current object and its children. For example, a simple search at the root DataQuery level would search over all datasets (name, description, etc) and all of the fields in the PIF. If the simple search field is filled out on say a ProcessStep object, then it will execute against the name of the process step and all of the nested objects.