2.5.0 Sorting support
This adds support for sorting in two places. In the first, a use can add to any FieldQuery-type object a sort order - either ASCENDING or DESCENDING. This will allow for sorting on any field within a PIF. In the second, a user can specify at the top level of a query the extractAs key and sort order if sorted on an extracted value. This is a convenience method that can be used to sort dynamically based on extraction keys, rather than having to inject into a specific localtion in the PIF.