Refactor field queries
In order to make a place for Analysis queries against any field, we need to wrap Filter objects in FieldQuery-like objects that can save the Analsys to run. This commit contains a refactor of FieldQuery hierarchy. No fields get a list of Filter objects anymore - instead, each gets something that is derived from AbstractFieldQuery, which in a later commit will be used to wrap the Analysis object.

To maintain backwards compatibility, this also introduces a deserializer to handle conversion from serialized Filter objects to BasicFieldQuery objects. It also defines a ConvertsToBasic(Boolean)FieldQuery interface that marks Filter and BasicFieldQuery (BooleanFilter and BasicBooleanFieldQuery) objects for backwards compatability in code.